+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ### Changelog
 
+**0.86.7**
+
+* Project Editor - Timer delay duration has been updated
+* "angularx-qrcode" package has been added - https://github.com/cordobo/angularx-qrcode
+  * QRCodeJs library has been changed with "angularx-qrcode" library
+  * (QRCodeJs library has been removed)
+
 **0.86.6**
 
 * Timer related calculation fixes